# cmp: optimize Compare