Compare provides three-way comparison for Ordered types, which certain
algorithms benefit from.

Float comparisons require NaN handling that should be optimized away for
all other types.
String comparisons benefit from using != which short-circuits on
different length strings.
Ideally the compiler would recognize the pattern and specialize on
strings, but this tries to improve the status quo.

goos: darwin
goarch: amd64
pkg: cmp
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
                   │  base.txt   │              patch.txt              │
                   │   sec/op    │   sec/op     vs base                │
Compare_int-12       1.638n ± 1%   1.635n ± 2%        ~ (p=0.698 n=10)
Compare_float64-12   2.397n ± 1%   2.206n ± 0%   -7.99% (p=0.000 n=10)
Compare_strings-12   8.054n ± 1%   6.705n ± 2%  -16.75% (p=0.000 n=10)
geomean              3.162n        2.892n        -8.56%

Updates #31187
Updates #61725