@ncruces 2024-10-14
https://go-review.googlesource.com/c/go/+/578835